updated index.js
decided to add all my inquirer prompts here instead of defining classes in a lib directory. Added chalk for styling our CLI output, and console.log to display the data in table view. I added a mainMenu inquirer prompt, and if statements to run functions if an option is selected.